refactor(anthropic): extract request conversion
